ld_conf.rs: Module to read ld.so.conf definitions
The ld.so.conf format is briefly defined by ldconfig manpage [1].
It allows specify system directories to be part of the library
resolution, the ldconfig command will read the file and any other
specified with the 'include' directive and create a binary cache
(/etc/ld.so.cache).  The cache is subject to change by the loader.

It currently only support absolute path in the 'define' directly
and it also does not support the expected type (dirname=TYPE)
which is deprecated by glibc.

[1] https://man7.org/linux/man-pages/man8/ldconfig.8.html