main.rs: Use canonicalize on the executable filename
On glibc/Linux the RTLD_DI_ORIGIN for the executable itself (used
for $ORIGIN expansion) is obtained by first following the]
'/proc/self/exe' symlink and if it is not available the loader also
checks the 'LD_ORIGIN_PATH' environment variable.

The '/proc/self/exec' is an absolute path and to mimic loader
behavior we first try to canocalize the input filename to remove any
symlinks.  There is not much sense in trying LD_ORIGIN_PATH, since it
is is only checked by the loader if the binary can not dereference
the procfs entry.