macho: Add support to consult the dyld cache
macOS starting with BigSur only provides a generated cache of all built
dynamic libraries, so the library file does not exist in the file system.

So to get all the cache hit dependendy, it requires to locale and parse
the object within the dyld cache.  To avoid overhead, the image cache
now maps the library path used on LC_DYLIB to a file offset.  Once the
dependency is consulted, the cahce is parse and the library dependencies
is returned.