elf: Parse ld.so.cache instead of ld.so.conf
The ld.so.cache is the de facto cache used by the system, the ld.so.conf
might be no synced with system one and the ld.so.cache contains more
information than the ld.so.conf (such as the new glibc-hwcap).

Currently only the newest format (glibc-ld.so.cache1.1) is supported
and the symbol search list if constructed without filtering or enable
hwcaps.