elf: Optimize ld.so.cache read
Every BufReader seelk operation always discard the internal buf, and
then it triggers a new read syscall (unfortunately rust runtime does
not use pread or similar syscall to atomically read a file with a
define position).

However BufReader does provide seek_relative, which does not discard
the buffer if the seek position lies within the buffer.  The drawback
is it requires to keep track of the current position.