macho: Add DYLD_LIBRARY_PATH support
It requires to add an extra handling on arenatree::EqualString so
entries that are LdLibraryPath is compare against only the input
name.  This is to avoid a circular dependency on some libraries
that referennces itself and since the deptree entry is the one
with DYLD_LIBRARY_PATH resolved it does not fully match.

Also rename --ld-library-path to just --library-path to avoid confusion
between elf and macho.