elf: Add support to glibc-hwcap
The glibc 2.33 added support for glibc-hwcap loader extension [1], which
allows glibc loader to search for additional implicit directories for
optimized libraries according to a pre-defined ABIs sets.

For instance, on x86_64 three abis levels are defined: x86-64-v2, x86-64-v3,
x86-64-v4.  Assuming the folder glibc-hwcaps exists in any searched
folder for the ld.so.conf, any library within the abi defined folder are
also searched.  So for the following library disposition:

  /lib64/glibc-hwcaps/x86-64-v3/libfoo.1
  /lib64/glibc-hwcaps/x86-64-v2/libfoo.1
  /lib64/libfoo.1

The ldconfig will create ld.so.cache entries for each library with
additional metadata to indicate that libfoo.1 has ABI support for both
x86-64-v3 and x86-64-v2.

At loading time, ld.so will check current CPU support and match against
the provided ld.so.cache libraries.  So, if the CPU has support for
x86-64-v3, /lib64/glibc-hwcaps/x86-64-v3/libfoo.1 will be loader.  If
the cpu has only support for x86-64-v2, the
/lib64/glibc-hwcaps/x86-64-v2/libfoo.1 will be selected.

To support it on rldd it requires additional support to query current
CPU support. It is done through cpuid instruction on x86_64 and parsing
/proc/self/auxv on PowerPC and s390x.

The ld.so.cache is then parsed and only the best-fit is added on
ldsocache.

[1] https://sourceware.org/pipermail/libc-alpha/2020-October/118184.html