elf: android: Check if namespace is accessible
And if the namespace is not isolated also check for the allowed libs.
Also remove the permitted paths, since it is not  prepended to
the requested library names when the dynamic linker is searching for a
shared library.

It is used to allow programers to load shared libraries into an
isolated namespace by specifying the full path of the shared
library, and since r23 DT_NEEDED with full path is considered as
invalid ones.